requirements.txt
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.7
mypy==1.10.0
mypy-extensions==1.0.0
nba_api==1.4.1
numpy==1.26.4
pandas==2.2.2
pandas-stubs==2.2.2.240514
praw==7.7.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.6
tomli==2.0.1
types-pytz==2024.1.0.20240417
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
